LayoutTests/storage/indexeddb/index-get-key-argument-required-expected.txt

Test IndexedDB index .get() required argument

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;
PASS indexedDB == null is false
IDBDatabaseException = window.IDBDatabaseException || window.webkitIDBDatabaseException;
PASS IDBDatabaseException == null is false
IDBCursor = window.IDBCursor || window.webkitIDBCursor;
PASS IDBCursor == null is false
IDBKeyRange = window.IDBKeyRange || window.webkitIDBKeyRange;
PASS IDBKeyRange == null is false
indexedDB.open(name, description)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo', { keyPath: 'id', autoIncrement: true });
index = objectStore.createIndex('first', 'first');
PASS index.get(); threw exception TypeError: Not enough arguments.
PASS index.getKey(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
